requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
cffi==1.12.2
cryptography==2.6.1
future==0.17.1
more-itertools==6.0.0
pluggy==0.9.0
py==1.8.0
pycparser==2.19
PySocks==1.6.8
pytest==4.3.0
python-telegram-bot==11.1.0
six==1.12.0
SQLAlchemy==1.2.18
